enable setting a double ratio, rather that just int (#21)

+++ spring-cloud-sleuth-otel-autoconfigure/src/main/java/org/springframework/cloud/sleuth/autoconfig/otel/OtelProperties.java
@@ -95,7 +95,7 @@ public class OtelProperties {
 		return this.traceIdRatioBased;
 	}
 
-	public void setTraceIdRatioBased(int traceIdRatioBased) {
+	public void setTraceIdRatioBased(double traceIdRatioBased) {
 		this.traceIdRatioBased = traceIdRatioBased;
 	}
